+++ src/_modules/root/provinces.ts
@@ -1,0 +1,166 @@
+export const zambiaProvinces = [
+    {
+      name: "Central Province",
+      capital: "Kabwe",
+      image: "central_province.webp",
+      districts: [
+        { name: "Kabwe" },
+        { name: "Chibombo" },
+        { name: "Kapiri Mposhi" },
+        { name: "Mkushi" },
+        { name: "Serenje" },
+        { name: "Chisamba" },
+        { name: "Ngabwe" },
+        { name: "Luano" }
+      ]
+    },
+    {
+      name: "Copperbelt Province",
+      capital: "Ndola",
+      image: "copperbelt_province.webp",
+      districts: [
+        { name: "Ndola" },
+        { name: "Kitwe" },
+        { name: "Chingola" },
+        { name: "Mufulira" },
+        { name: "Luanshya" },
+        { name: "Kalulushi" },
+        { name: "Chililabombwe" },
+        { name: "Lufwanyama" },
+        { name: "Masaiti" },
+        { name: "Mpongwe" }
+      ]
+    },
+    {
+      name: "Eastern Province",
+      capital: "Chipata",
+      image: "eastern_province.webp",
+      districts: [
+        { name: "Chipata" },
+        { name: "Petauke" },
+        { name: "Katete" },
+        { name: "Lundazi" },
+        { name: "Sinda" },
+        { name: "Chadiza" },
+        { name: "Nyimba" },
+        { name: "Mambwe" },
+        { name: "Vubwi" }
+      ]
+    },
+    {
+      name: "Luapula Province",
+      capital: "Mansa",
+      image: "luapula_province.webp",
+      districts: [
+        { name: "Mansa" },
+        { name: "Samfya" },
+        { name: "Nchelenge" },
+        { name: "Mwense" },
+        { name: "Kawambwa" },
+        { name: "Chiengi" },
+        { name: "Chembe" },
+        { name: "Milenge" },
+        { name: "Mwansabombwe" },
+        { name: "Chipili" },
+        { name: "Lunga" }
+      ]
+    },
+    {
+      name: "Lusaka Province",
+      capital: "Lusaka",
+      image: "lusaka_province.webp",
+      districts: [
+        { name: "Lusaka" },
+        { name: "Chongwe" },
+        { name: "Kafue" },
+        { name: "Chilanga" },
+        { name: "Luangwa" },
+        { name: "Rufunsa" }
+      ]
+    },
+    {
+      name: "Muchinga Province",
+      capital: "Chinsali",
+      image: "muchinga_province.webp",
+      districts: [
+        { name: "Chinsali" },
+        { name: "Mpika" },
+        { name: "Isoka" },
+        { name: "Nakonde" },
+        { name: "Shiwang'andu" },
+        { name: "Kanchibiya" },
+        { name: "Lavushimanda" },
+        { name: "Chama" }
+      ]
+    },
+    {
+      name: "Northern Province",
+      capital: "Kasama",
+      image: "northern_province.webp",
+      districts: [
+        { name: "Kasama" },
+        { name: "Mbala" },
+        { name: "Mpulungu" },
+        { name: "Mungwi" },
+        { name: "Luwingu" },
+        { name: "Mporokoso" },
+        { name: "Kaputa" },
+        { name: "Chilubi" },
+        { name: "Senga Hill" },
+        { name: "Lupososhi" }
+      ]
+    },
+    {
+      name: "North-Western Province",
+      capital: "Solwezi",
+      image: "north_western_province.webp",
+      districts: [
+        { name: "Solwezi" },
+        { name: "Kasempa" },
+        { name: "Mufumbwe" },
+        { name: "Kabompo" },
+        { name: "Zambezi" },
+        { name: "Chavuma" },
+        { name: "Manyinga" },
+        { name: "Mushindamo" },
+        { name: "Kalumbila" }
+      ]
+    },
+    {
+      name: "Southern Province",
+      capital: "Choma",
+      image: "southern_province.webp",
+      districts: [
+        { name: "Choma" },
+        { name: "Livingstone" },
+        { name: "Mazabuka" },
+        { name: "Monze" },
+        { name: "Kalomo" },
+        { name: "Siavonga" },
+        { name: "Gwembe" },
+        { name: "Sinazongwe" },
+        { name: "Kazungula" },
+        { name: "Pemba" },
+        { name: "Zimba" }
+      ]
+    },
+    {
+      name: "Western Province",
+      capital: "Mongu",
+      image: "western_province.webp",
+      districts: [
+        { name: "Mongu" },
+        { name: "Senanga" },
+        { name: "Kalabo" },
+        { name: "Shangombo" },
+        { name: "Kaoma" },
+        { name: "Lukulu" },
+        { name: "Sesheke" },
+        { name: "Nalolo" },
+        { name: "Limulunga" },
+        { name: "Mwandi" },
+        { name: "Sioma" }
+      ]
+    }
+  ];
+